+++ spec/swagger_v2/api_swagger_v2_headers_spec.rb
@@ -18,6 +18,9 @@ describe 'headers' do
              },
 
              entity: Entities::UseResponse
+        params do
+          optional :param_x, type: String, desc: 'This is a parameter', documentation: { param_type: 'query' }
+        end
         get '/use_headers' do
           { 'declared_params' => declared(params) }
         end
@@ -37,15 +40,17 @@ describe 'headers' do
   end
 
   specify do
-    expect(subject['paths']['/use_headers']['get']['parameters']).to eql(
-      [
-        { 'in' => 'header',
-          'name' => 'X-Rate-Limit-Limit',
-          'description' => 'The number of allowed requests in the current period',
-          'type' => 'integer',
-          'format' => 'int32',
-          'required' => false }
-      ]
+    parameters = subject['paths']['/use_headers']['get']['parameters']
+    expect(parameters).to include(
+      'in' => 'header',
+      'name' => 'X-Rate-Limit-Limit',
+      'description' => 'The number of allowed requests in the current period',
+      'type' => 'integer',
+      'format' => 'int32',
+      'required' => false
     )
+    expect(parameters.size).to eq(2)
+    expect(parameters.first['in']).to eq('header')
+    expect(parameters.last['in']).to eq('query')
   end
 end